Clarity: rework to use latest microscope version.
It is better to use MicroscopeBase instead of MicroscopeFilter, because the microscope implementation is not based on FilterWheelBase anymore and therefore using MicroscopeFilter would require extra effort and hacks. Using MicroscopeBase directly also leaves the handling of the filter turret position to the custom UI panel instead of making a generic filter selector together with the other filters. The benefit of this is that the filter choice widget can be disabled/enabled together with the Clarity device. The reworked class also includes extra functionality.